requirements.txt
SQLAlchemy==2.0.40
fake-useragent==2.2.0
requests==2.32.3
beautifulsoup4==4.13.4
pydantic==2.11.4
uvicorn==0.34.2
more-itertools==10.7.0
python-dotenv~=1.0.1
fastapi~=0.115.8
alembic==1.15.2
psycopg2-binary